refactor: simplify focus query layer — merge archive fns, NOT EXISTS, dedupe test helpers, fix FK on entry_focus
